### 2017-06-04 #275 @statsmaths
Here are a few people who would be good reviewers but were not involved with the package itself: Mike Kane (@kaneplusplus), Akitaka Matsuo (@amatsuo), Adam Obeng (@adamobeng).
