### 2017-06-14 #296 @whedon
[REVIEW]: kerasR: R Interface to the Keras Deep Learning Library
**Submitting author:** @statsmaths (<a href="http://orcid.org/0000-0003-0576-0669">Taylor Arnold</a>)
**Repository:** <a href="https://github.com/statsmaths/kerasR" target ="_blank">https://github.com/statsmaths/kerasR</a>
**Version:** 0.6.1
