### 2017-06-19 #296 @kaneplusplus
This is a package I have begun using when applying deep learning models. This interface mirrors Python's making it very easy to transition between the two languages. Constructing a model, fitting it, and predicting is easy. The examples in the documentation were helpful and I liked that the vignette includes references for other deep learning resources.

The package should be accepted. While going through the code and documentation, I came up with the following three suggestions.

- (suggestion) The installation assumes that keras is installed and configured for python. Is it worth pointing to installation instructions or adding the `pip install` commands that can be used to set this up? This would include installing `pydot` for visualization.
- (suggestion) There is code output in [the vignette](https://github.com/statsmaths/kerasR/blob/master/vignettes/introduction.Rmd) that simply reads `output`. Should this be removed?
- (suggestion) The code coverage is slightly low. Will this be increased in the coming weeks?
