### 2017-06-19 #296 @statsmaths
@kaneplusplus Thank you very much for reviewing this submission and your very helpful feedback. I just pushed version 0.7.0 to GitHub, which I believe implements your three suggestions. Namely:

- The package now gives an explicit warning message whenever any keras-related function is called if the keras package is not available. Details instructions for using pip to install the requirements are now given. I've also made it so that there is an explicit printing of the python path being used by reticulate.
- I removed the erroneous `output` line in the vignette and README file.
- Due to an oversight on my part, many of the test files were not pushed to the repository. These are now online. Because CRAN does not have the Python keras module, and some of the tests take a while to run, there are `skip_on_cran()` calls in the testing suite, so you need to set the `export NOT_CRAN='true'` if you want to run these locally. The current coverage is now at 76%.

Please let me know if you have any further questions or suggestions.
