### 2017-06-20 #296 @karthik
@kaneplusplus Thank you very much for the review. We really appreciate it.

@statsmaths Could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
