### 2017-06-20 #296 @statsmaths
@karthik Thanks. I archived the current version of the software to Zenodo with this DOI: 10.5281/zenodo.814996
