### 2017-06-22 #296 @arfon
@kaneplusplus many thanks for your review here and to @karthik for editing this one :sparkles:

@statsmaths - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00296 :zap: 🚀:boom:
